requirements.txt
python-dotenv
pandas 
numpy
pymongo
certifi
pymongo[srv]==3.6
scikit-learn

-e .